chore: command/job test layout: move inline tests to existing tests/ di… [ORB-10388]
command/job test layout: move inline tests to existing tests/ directory

Planned-By: codex
Agent-Model: gpt-5.6-terra
