+++ crates/orbit-core/src/command/job/tests/mod.rs
@@ -1,1 +1,2 @@
 mod catalog;
+mod exec;
